# api-rest-fastapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Taller 02 – API REST de Envíos (FastAPI)",
    "version": "1.0.0",
    "description": "API para gestión de envíos de una empresa logística. Implementa endpoints básicos, almacenamiento en memoria y documentación OpenAPI."
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ]
                },
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        }
      }
    },
    "/envios": {
      "get": {
        "summary": "Lista todos los envíos",
        "operationId": "listar_envios_envios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Envio"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crea un nuevo envío",
        "operationId": "crear_envio_envios_post",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envio"
                }
              }
            }
          },
          "409": {
            "description": "ID ya existe"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Envio"
              }
            }
          }
        }
      }
    },
    "/envios/{envio_id}": {
      "get": {
        "summary": "Obtiene un envío por ID",
        "operationId": "obtener_envio_envios__envio_id__get",
        "parameters": [
          {
            "name": "envio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envio"
                }
              }
            }
          },
          "404": {
            "description": "Envío no encontrado"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Envio": {
        "title": "Envio",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identificador único del envío"
          },
          "destinatario": {
            "type": "string",
            "title": "Destinatario",
            "description": "Nombre del destinatario"
          },
          "direccion": {
            "type": "string",
            "title": "Direccion",
            "description": "Dirección de entrega"
          },
          "estado": {
            "type": "string",
            "title": "Estado",
            "description": "Estado actual del envío"
          }
        },
        "required": [
          "id",
          "destinatario",
          "direccion",
          "estado"
        ],
        "examples": [
          {
            "id": "001",
            "destinatario": "Juan Pérez",
            "direccion": "Av. Siempre Viva 742",
            "estado": "En tránsito"
          }
        ]
      }
    }
  }
}